Adding "exec" to the run.sh
Calibre-Web does not receive the SIGTERM signal from docker when attempting to stop the container. This causes docker to kill the container after 10 seconds and the server is not shut down gracefully. Adding "exec" causes the python command to become PID 1 instead of run.sh, then the SIGTERM signal is sent directly to the python process.